+++ client/src/components/authentication/LoginForm.tsx
@@ -1,14 +1,10 @@
-import { useMutation } from '@tanstack/react-query';
-import { AxiosError } from 'axios';
 import { useState } from 'react';
 import { BsEyeFill, BsEyeSlashFill } from 'react-icons/bs';
 import { HiXCircle } from 'react-icons/hi';
-import { useNavigate } from 'react-router-dom';
-import { Login } from '@/api/api';
+
 import Button from '@/components/@common/button/Button';
-import { REFRSH_TOKEN_DURATION } from '@/constant/constantValue';
-import { LoginInfo } from '@/types/types';
-import { validateTokens } from '@/utils/validateTokens';
+
+import useMemberLogin from '@/hooks/useMemberLogin';
 
 function LoginForm() {
   const [email, setEmail] = useState('');
@@ -17,8 +13,6 @@ function LoginForm() {
 
   const [emailError, setEmailError] = useState<string | null>(null);
   const [passwordError, setPasswordError] = useState<string | null>(null);
-  const [loginError, setLoginError] = useState<string | null>(null);
-  const navigate = useNavigate();
 
   const handleEmail = (e: React.ChangeEvent<HTMLInputElement>) => {
     setEmail(e.target.value);
@@ -53,29 +47,8 @@ function LoginForm() {
   function validate() {
     return checkEmail() && checkuserPassword();
   }
+  const { mutation, loginError } = useMemberLogin();
 
-  const mutation = useMutation({
-    mutationFn: (member: LoginInfo) => Login(member),
-    onSuccess(data) {
-      if (data.status === 200) {
-        validateTokens(data.headers.authorization, data.headers.refresh);
-        const expiration = new Date();
-        expiration.setMinutes(expiration.getMinutes() + REFRSH_TOKEN_DURATION);
-        localStorage.setItem('expiration', expiration.toISOString());
-        navigate('/');
-      }
-    },
-    onError(error: AxiosError) {
-      if (error.response && error.response.status === 401) {
-        setLoginError(
-          '로그인에 실패했습니다. 이메일과 비밀번호를 확인해주세요.'
-        );
-        console.error('error:  with same data already exist.');
-      } else {
-        console.error('error:', error);
-      }
-    },
-  });
   const submitHandler = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     if (validate()) {
