+++ client/src/components/comments/CommentForm.tsx
@@ -4,7 +4,7 @@ import { BiPaperPlane } from 'react-icons/bi';
 import { useParams } from 'react-router-dom';
 import { styled } from 'styled-components';
 import { PostComment } from '@/api/api';
-import checkLogin from '@/utils/isLoggedIn';
+import checkLogin from '@/utils/checkLogin';
 
 function CommentForm() {
   const [content, setContent] = useState('');
