+++ client/src/components/mediaDetail/recommend/Recommend.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from 'react-router-dom';
 import { GetIsRecommend, PostRecommend } from '@/api/api';
 import RecommendLoading from '@/components/mediaDetail/recommend/RecommendLoading';
 import { S_IconWrapper } from '@/styles/style';
-import checkLogin from '@/utils/isLoggedIn';
+import checkLogin from '@/utils/checkLogin';
 import { notifyError, notifyWithIcon } from '@/utils/notify';
 
 function Recommend({
