+++ client/src/constant/constantValue.ts
@@ -3,7 +3,7 @@ export const PAGES_PER_SECTION = 5;
 export const ADMIN_MEMBERID = 2;
 export const AUTOCOMPLETE_RESULT_SIZE = 5;
 export const SCROLL_Y_SECTION_1 = 400;
-export const REFRSH_TOKEN_DURATION = 1440;
+export const REFRESH_TOKEN_DURATION = 1440;
 
 export const genres = [
   '액션',
